refactor: renderer
