+++ src/renderer/index.ts
@@ -1,0 +1,3 @@
+export { renderHeader } from './header.js'
+export { renderUsage } from './usage.js'
+export { renderValidationErrors } from './validation.js'
